Make YAML tests run for single node
