+++ x-pack/plugin/esql/qa/server/single-node/src/yamlRestTest/java/org/elasticsearch/xpack/esql/qa/single_node/AbstractEsqlClientYamlIT.java
@@ -11,12 +11,22 @@ import org.elasticsearch.test.cluster.ElasticsearchCluster;
 import org.elasticsearch.test.cluster.local.distribution.DistributionType;
 import org.elasticsearch.test.rest.yaml.ClientYamlTestCandidate;
 import org.elasticsearch.test.rest.yaml.ESClientYamlSuiteTestCase;
+import org.elasticsearch.test.rest.yaml.section.ApiCallSection;
+import org.elasticsearch.test.rest.yaml.section.ClientYamlTestSection;
+import org.elasticsearch.test.rest.yaml.section.DoSection;
+import org.elasticsearch.test.rest.yaml.section.ExecutableSection;
 import org.elasticsearch.xpack.esql.qa.rest.EsqlSpecTestCase;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+
 abstract class AbstractEsqlClientYamlIT extends ESClientYamlSuiteTestCase {
+    static String ESQL_VERSION = "2024.04.01.🚀";
 
     @ClassRule
     public static ElasticsearchCluster cluster = ElasticsearchCluster.local()
@@ -43,5 +53,50 @@ abstract class AbstractEsqlClientYamlIT extends ESClientYamlSuiteTestCase {
          * it private works - the hook still runs. It just looks strange.
          */
         EsqlSpecTestCase.assertRequestBreakerEmpty();
+    }
+
+    public static Iterable<Object[]> updateEsqlQueryDoSections(Iterable<Object[]> parameters, Function<DoSection, ExecutableSection> modify)
+        throws Exception {
+        List<Object[]> result = new ArrayList<>();
+        for (Object[] orig : parameters) {
+            assert orig.length == 1;
+            ClientYamlTestCandidate candidate = (ClientYamlTestCandidate) orig[0];
+            try {
+                ClientYamlTestSection modified = new ClientYamlTestSection(
+                    candidate.getTestSection().getLocation(),
+                    candidate.getTestSection().getName(),
+                    candidate.getTestSection().getPrerequisiteSection(),
+                    candidate.getTestSection().getExecutableSections().stream().map(e -> modifyExecutableSection(e, modify)).toList()
+                );
+                result.add(new Object[] { new ClientYamlTestCandidate(candidate.getRestTestSuite(), modified) });
+            } catch (IllegalArgumentException e) {
+                throw new IllegalArgumentException("error modifying " + candidate + ": " + e.getMessage(), e);
+            }
+        }
+        return result;
+    }
+
+    private static ExecutableSection modifyExecutableSection(ExecutableSection e, Function<DoSection, ExecutableSection> modify) {
+        if (false == (e instanceof DoSection)) {
+            return e;
+        }
+        DoSection doSection = (DoSection) e;
+        String api = doSection.getApiCallSection().getApi();
+        return switch (api) {
+            case "esql.query" -> modify.apply(doSection);
+            case "esql.async_query", "esql.async_query_get" -> throw new IllegalArgumentException(
+                "The esql yaml tests can't contain async_query or async_query_get because we modify them on the fly and *add* those."
+            );
+            default -> e;
+        };
+    }
+
+    static DoSection setVersion(DoSection doSection) {
+        ApiCallSection copy = doSection.getApiCallSection().copyWithNewApi(doSection.getApiCallSection().getApi());
+        for (Map<String, Object> body : copy.getBodies()) {
+            body.putIfAbsent("version", ESQL_VERSION);
+        }
+        doSection.setApiCallSection(copy);
+        return doSection;
     }
 }
