+++ x-pack/plugin/esql/qa/server/single-node/src/yamlRestTest/java/org/elasticsearch/xpack/esql/qa/single_node/EsqlClientYamlAsyncSubmitAndFetchIT.java
@@ -32,7 +32,7 @@ public class EsqlClientYamlAsyncSubmitAndFetchIT extends AbstractEsqlClientYamlI
 
     @ParametersFactory
     public static Iterable<Object[]> parameters() throws Exception {
-        return EsqlClientYamlAsyncIT.parameters(DoEsqlAsync::new);
+        return updateEsqlQueryDoSections(createParameters(), doSection -> { return new DoEsqlAsync(setVersion(doSection)); });
     }
 
     private static class DoEsqlAsync implements ExecutableSection {
